Split main into commands
